requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
djangorestframework==3.14.0
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
outcome==1.2.0
packaging==23.1
psycopg2-binary==2.9.7
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.10.4
urllib3==2.0.4
webdriver-manager==4.0.0
wsproto==1.2.0
